[Decomp][Player] Cohesion split: extract gossip system to PlayerGossip.cpp
Moves the GOSSIP SYSTEM banner and all gossip-menu methods out of
Player.cpp into a new file under src/game/Object/.

Player.cpp: 17,700 -> 17,170 lines (-530).
PlayerGossip.cpp: 613 lines, 6 methods.

Methods moved:
- PrepareGossipMenu, SendPreparedGossip, OnGossipSelect
- GetGossipTextId (both overloads)
- GetDefaultGossipMenuForSource

No header change; no signature change; no behaviour change.
